==> 05_graphics_tables.Rmd <==
---
title: "graphics_tables"
author: "Leandro Pereira Garcia"
date: "November, 2020"
output: 
  word_document: 
    #reference_docx: paper.docx
    fig_caption: yes
    fig_height: 8
    fig_width: 12
    toc: no
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning=FALSE)
options(scipen=999)
set.seed(233)
```


```{r include=FALSE}
#Libraries
library(readr)
library(tidyverse)
library(ggplot2)
library(sf)
library(spData)
library(ggpubr)
library(RColorBrewer)
library(corrplot)
library(ggcorrplot)
library(reshape2)
```


```{r include=FALSE}
#Databases
base_efi <- read_csv("bases/completed_base.csv")
impact <- read_csv("bases/impact.csv")
impact_sensib_imputation <- read_csv("bases/impact_sensib_imputation.csv")
impact_sensib_gps_model <- read_csv("bases/impact_sensib_gps_model.csv")

```

```{r echo=TRUE}
#Description
base_efi[which(base_efi$MEAN_RATE_NEO == min(base_efi$MEAN_RATE_NEO, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "MEAN_RATE_NEO")]
base_efi[which(base_efi$MEAN_RATE_NEO == max(base_efi$MEAN_RATE_NEO, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "MEAN_RATE_NEO")]
```

```{r echo=TRUE}
base_efi[which(base_efi$MEAN_RATE_NEO_U5 == min(base_efi$MEAN_RATE_NEO_U5, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "MEAN_RATE_NEO_U5")]
base_efi[which(base_efi$MEAN_RATE_NEO_U5 == max(base_efi$MEAN_RATE_NEO_U5, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "MEAN_RATE_NEO_U5")]
```

```{r echo=TRUE}
base_efi[which(base_efi$PUBLIC_EXP_PER_CAP_LAGGED == min(base_efi$PUBLIC_EXP_PER_CAP_LAGGED, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "PUBLIC_EXP_PER_CAP_LAGGED")]
base_efi[which(base_efi$PUBLIC_EXP_PER_CAP_LAGGED == max(base_efi$PUBLIC_EXP_PER_CAP_LAGGED, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "PUBLIC_EXP_PER_CAP_LAGGED")]
```

```{r echo=TRUE}
base_efi[which(base_efi$HEALTH_EXP_LAGGED == min(base_efi$HEALTH_EXP_LAGGED, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "HEALTH_EXP_LAGGED")]
base_efi[which(base_efi$HEALTH_EXP_LAGGED == max(base_efi$HEALTH_EXP_LAGGED, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "HEALTH_EXP_LAGGED")]
```

```{r echo=TRUE}
base_efi[which(base_efi$OTHER_EXP_LAGGED == min(base_efi$OTHER_EXP_LAGGED, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "OTHER_EXP_LAGGED")]
base_efi[which(base_efi$OTHER_EXP_LAGGED == max(base_efi$OTHER_EXP_LAGGED, na.rm = T)), colnames(base_efi) %in% c("LOCATION", "OTHER_EXP_LAGGED")]
```



```{r include=FALSE}
#Importing the world map
WorldData <- map_data('world')
countries_world <- data.frame(WORLD = WorldData$region) 
WorldData %>% filter(region != "Antarctica") -> WorldData
WorldData <- fortify(WorldData)


base_efi_loc <- data.frame(BASE_EFI_COUNTRIES = unique(base_efi$LOCATION),
			    BASE_EFI_COUNTRIES_2 = unique(base_efi$LOCATION))
	
countries <- merge(countries_world, base_efi_loc, by.x = "WORLD", by.y = "BASE_EFI_COUNTRIES_2", all = T) %>% unique()


base_efi[which(base_efi$LOCATION == "Bolivia (Plurinational State of)"), 1] <- "Bolivia"
base_efi[which(base_efi$LOCATION == "Czechia"), 1] <- "Czech Republic"
base_efi[which(base_efi$LOCATION == "Iran (Islamic Republic of)"), 1] <- "Iran"
base_efi[which(base_efi$LOCATION == "Lao People's Democratic Republic"), 1] <- "Laos"
base_efi[which(base_efi$LOCATION == "Democratic People's Republic of Korea"), 1] <- "North Korea"
base_efi[which(base_efi$LOCATION == "Congo"), 1] <- "Republic of Congo"
base_efi[which(base_efi$LOCATION == "Russian Federation"), 1] <- "Russia"
base_efi[which(base_efi$LOCATION == "United Republic of Tanzania"), 1] <- "Tanzania"
base_efi[which(base_efi$LOCATION == "United Kingdom"), 1] <- "UK"
base_efi[which(base_efi$LOCATION == "United States of America"), 1] <- "USA"
base_efi[which(base_efi$LOCATION == "Venezuela (Bolivarian Republic of)"), 1] <- "Venezuela"
base_efi[which(base_efi$LOCATION == "Viet Nam"), 1] <- "Vietnam"
base_efi[which(base_efi$LOCATION == ""), 1] <- ""


base_map <- base_efi
base_map <- merge(base_map, WorldData, by.x = "LOCATION", by.y = "region", all.x = T)

```

```{r include=FALSE}

myPalette_rev <- colorRampPalette(brewer.pal(11, "Spectral")) #Increasing
myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")))  #Decreasing

```



```{r fig.width=15, fig.height=6}
#Mortality

plot1 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_map, map=WorldData,
                  aes(fill= MEAN_RATE_NEO, map_id=LOCATION),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Mean NeoRt 2018-2019", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette(20), 
                             labels = round(quantile(base_map$MEAN_RATE_NEO, c(.05,.95)),2), 
                             breaks = round(quantile(base_map$MEAN_RATE_NEO, c(.05,.95)),2))



plot2 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_map, map=WorldData,
                  aes(fill= MEAN_RATE_NEO_U5, map_id=LOCATION),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Mean NeoU5Rt 2018-2019", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette(20), 
                             labels = round(quantile(base_map$MEAN_RATE_NEO_U5, c(.05,.95)),2), 
                             breaks = round(quantile(base_map$MEAN_RATE_NEO_U5, c(.05,.95)),2))


ggarrange(plot1, plot2, labels = c("A", "B"), ncol = 2,nrow = 1,
           common.legend = F,legend = "bottom")

```


```{r fig.width=16, fig.height=10}
#Treatment

plot1 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_map, map=WorldData,
                  aes(fill= LOG_PUBLIC_EXP_PER_CAP_LAGGED, map_id=LOCATION),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Log of Mean Total Public Exp 2013-2017", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette_rev(20), 
                             labels = round(quantile(base_map$LOG_PUBLIC_EXP_PER_CAP_LAGGED, c(.05,.95)),2), 
                             breaks = round(quantile(base_map$LOG_PUBLIC_EXP_PER_CAP_LAGGED, c(.05,.95)),2))



plot2 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_map, map=WorldData,
                  aes(fill= LOG_HEALTH_EXP_LAGGED, map_id=LOCATION),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Log of Mean Health Public Exp 2013-2017", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette_rev(20), 
                             labels = round(quantile(base_map$LOG_HEALTH_EXP_LAGGED, c(.05,.95)),2), 
                             breaks = round(quantile(base_map$LOG_HEALTH_EXP_LAGGED, c(.05,.95)),2))

plot3 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_map, map=WorldData,
                  aes(fill= LOG_OTHER_EXP_LAGGED, map_id=LOCATION),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Log of Mean Other Sec Public Exp 2013-2017", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette_rev(20), 
                             labels = round(quantile(base_map$LOG_OTHER_EXP_LAGGED, c(.05,.95)),2), 
                             breaks = round(quantile(base_map$LOG_OTHER_EXP_LAGGED, c(.05,.95)),2))


plot4 <- ggarrange(plot2, plot3, labels = c("B", "C"),  ncol = 2, nrow = 1,common.legend = F, legend = "bottom")
ggarrange(plot1, plot4, labels = c("A"), nrow = 2, ncol = 1, common.legend = F, legend = "bottom")

```



```{r fig.width=12, fig.height=12}
impact_table <- impact
impact_table$Est. <- round(impact_table$Est., 3) 
impact_table$`0.5%` <- round(impact_table$`0.5%`, 3) 
impact_table$`99.5%` <- round(impact_table$`99.5%`, 3) 
impact_table$CI_99 <-  paste0("(",impact_table$`0.5%`,", ", impact_table$`99.5%`,")")
impact_table$`2.5%` <- round(impact_table$`2.5%`, 3) 
impact_table$`97.5%` <- round(impact_table$`97.5%`, 3) 
impact_table$CI_95 <-  paste0("(",impact_table$`2.5%`,", ", impact_table$`97.5%`,")")
impact_table$`0.1%` <- round(impact_table$`0.1%`, 3) 
impact_table$`100%` <- round(impact_table$`100%`, 3) 
impact_table$CI_999 <-  paste0("(",impact_table$`0.1%`,", ", impact_table$`100%`,")")
impact_table$"Impact\n(CI99%)\n(CI95%)\n(CI999%)" <- paste0(impact_table$Est.,impact_table$CI_99, impact_table$CI_95,impact_table$CI_999, sep="\n")
impact_table <- impact_table %>% dplyr::select(treatment, mortality_rate,"Impact\n(CI99%)\n(CI95%)\n(CI999%)")
names(impact_table) <- c("Treatment", "Mortality rate","Impact\n(CI99%)\n(CI95%)\n(CI999%)")
write.csv(impact_table, "bases/impact_table.csv", row.names = F)

```


```{r fig.width=12, fig.height=12}
impact_sensib_imputation_table <- impact_sensib_imputation
impact_sensib_imputation_table$Est. <- round(impact_sensib_imputation_table$Est., 3) 
impact_sensib_imputation_table$`0.5%` <- round(impact_sensib_imputation_table$`0.5%`, 3) 
impact_sensib_imputation_table$`99.5%` <- round(impact_sensib_imputation_table$`99.5%`, 3) 
impact_sensib_imputation_table$CI_99 <-  paste0("(",impact_sensib_imputation_table$`0.5%`,", ", impact_sensib_imputation_table$`99.5%`,")")
impact_sensib_imputation_table$`2.5%` <- round(impact_sensib_imputation_table$`2.5%`, 3) 
impact_sensib_imputation_table$`97.5%` <- round(impact_sensib_imputation_table$`97.5%`, 3) 
impact_sensib_imputation_table$CI_95 <-  paste0("(",impact_sensib_imputation_table$`2.5%`,", ", impact_sensib_imputation_table$`97.5%`,")")
impact_sensib_imputation_table$`0.1%` <- round(impact_sensib_imputation_table$`0.1%`, 3) 
impact_sensib_imputation_table$`100%` <- round(impact_sensib_imputation_table$`100%`, 3) 
impact_sensib_imputation_table$CI_999 <-  paste0("(",impact_sensib_imputation_table$`0.1%`,", ", impact_sensib_imputation_table$`100%`,")")
impact_sensib_imputation_table$"Impact\n(CI99%)\n(CI95%)\n(CI999%)" <- paste0(impact_sensib_imputation_table$Est.,impact_sensib_imputation_table$CI_99, impact_sensib_imputation_table$CI_95,impact_sensib_imputation_table$CI_999, sep="\n")
impact_sensib_imputation_table <- impact_sensib_imputation_table %>% dplyr::select(treatment, mortality_rate,"Impact\n(CI99%)\n(CI95%)\n(CI999%)")
names(impact_sensib_imputation_table) <- c("Treatment", "Mortality rate","Impact\n(CI99%)\n(CI95%)\n(CI999%)")
write.csv(impact_sensib_imputation_table, "bases/impact_sensib_imputation_table.csv", row.names = F)

```


```{r fig.width=12, fig.height=12}
impact_sensib_gps_model_table <- impact_sensib_gps_model
impact_sensib_gps_model_table$Est. <- round(impact_sensib_gps_model_table$Est., 3) 
impact_sensib_gps_model_table$`0.5%` <- round(impact_sensib_gps_model_table$`0.5%`, 3) 
impact_sensib_gps_model_table$`99.5%` <- round(impact_sensib_gps_model_table$`99.5%`, 3) 
impact_sensib_gps_model_table$CI_99 <-  paste0("(",impact_sensib_gps_model_table$`0.5%`,", ", impact_sensib_gps_model_table$`99.5%`,")")
impact_sensib_gps_model_table$`2.5%` <- round(impact_sensib_gps_model_table$`2.5%`, 3) 
impact_sensib_gps_model_table$`97.5%` <- round(impact_sensib_gps_model_table$`97.5%`, 3) 
impact_sensib_gps_model_table$CI_95 <-  paste0("(",impact_sensib_gps_model_table$`2.5%`,", ", impact_sensib_gps_model_table$`97.5%`,")")
impact_sensib_gps_model_table$`0.1%` <- round(impact_sensib_gps_model_table$`0.1%`, 3) 
impact_sensib_gps_model_table$`100%` <- round(impact_sensib_gps_model_table$`100%`, 3) 
impact_sensib_gps_model_table$CI_999 <-  paste0("(",impact_sensib_gps_model_table$`0.1%`,", ", impact_sensib_gps_model_table$`100%`,")")
impact_sensib_gps_model_table$"Impact\n(CI99%)\n(CI95%)\n(CI999%)" <- paste0(impact_sensib_gps_model_table$Est.,impact_sensib_gps_model_table$CI_99, impact_sensib_gps_model_table$CI_95,impact_sensib_gps_model_table$CI_999, sep="\n")
impact_sensib_gps_model_table <- impact_sensib_gps_model_table %>% dplyr::select(treatment, mortality_rate,"Impact\n(CI99%)\n(CI95%)\n(CI999%)")
names(impact_sensib_gps_model_table) <- c("Treatment", "Mortality rate","Impact\n(CI99%)\n(CI95%)\n(CI999%)")
write.csv(impact_sensib_gps_model_table, "bases/impact_sensib_gps_model_table.csv", row.names = F)

```
